Fix two silent data-loss bugs; add preflight, side verification, tests
Three faults each presented to the operator as "nothing works", with no
error and no diagnosis. All three are fixed, and the class is now guarded.

1. int16 overflow produced zero heel strikes on valid recordings

   The LSL outlet declares channel_format='int16', so pyxdf returns int16.
   compute_magnitude squared it: 4086^2 overflows int16, wraps negative, and
   sqrt of that is NaN. The NaN propagated through filtering and z-scoring,
   every threshold comparison returned False, and a clean 119s recording
   yielded 0 heel strikes -- surfaced only as "Cost extraction failed".
   Older files that happened to load as float32 masked this.

   compute_magnitude now casts to float64 before squaring, io.load_polar_stream
   loads float64 regardless of channel format, and compute_magnitude raises on
   non-finite or wrong-shaped input rather than returning a silent zero.

2. The H10 rejected the accelerometer configuration; nothing noticed

   ACC_WRITE carried a channel-count setting (0x04,0x01,0x03). The sensor
   answered f0 02 02 05 (ERROR INVALID PARAMETER) and sent nothing. The script
   never read the control response, so it printed success and left an LSL
   outlet open that carried no data. Verified by sweeping payload variants
   against a real device: identical request succeeds without that field.

   Field removed; the control point is now subscribed and its response
   checked, and sample flow is confirmed before the script reports ready.

3. A 1s LSL resolve window falsely reported disconnected sensors

   pylsl's default 1s multicast resolve intermittently misses healthy outlets
   on a shared lab network, and the console then reported SENSOR DISCONNECTED
   for sensors that never dropped a sample. Widened to 3s.

Also:

- cost.py: nine identical silent `return None` paths now record a specific
  reason (last_failure), and non-fatal QC warnings reach the caller instead of
  being discarded -- a trial that fell back to single-sensor mode previously
  looked identical to a clean one.
- console: shows the failure reason and any QC warnings.
- collect_sensors: --scan / --assign / --id. Sensor IDs move out of the source
  into config/sensors.json, since straps change between sessions.
- apps/verify_sides.py: guided shake test proving which LEG each stream is on.
  A device ID identifies a strap, not the limb it ended up on, and a silent L/R
  swap inverts the sign of the symmetry index.
- apps/preflight.py: one command run before a session -- tests, config, stale
  checkpoint, stream health, foreign streams, and analysis of the last
  recording.
- tests/test_regression.py: six tests encoding the above. No pytest required.
- Shebangs on the user-facing tools; `python3` on the lab machine resolves to an
  unrelated venv missing every dependency.

+++ hitlo/io.py
@@ -57,7 +57,9 @@ def load_polar_stream(xdf_path: str,
 
     for stream in data:
         if stream['info']['name'][0] == stream_name:
-            accel = np.asarray(stream['time_series'])
+            # float64 regardless of the stream's channel_format —
+            # int16 accel overflows as soon as anything squares it.
+            accel = np.asarray(stream['time_series'], dtype=np.float64)
             timestamps = np.asarray(stream['time_stamps'])
             if len(timestamps) < 2:
                 return None
